feature/UI-01: Battle UI has been started, added fonts to asset registry

+++ Kurumi-Engine/Engine/Systems/Rendering/Core/ActorRenderer.cs
@@ -41,8 +41,7 @@ public sealed class ActorRenderer
             var actor = currentActorRenderData.Actor;
 
             // Calculate the render layer.
-            var layer = actor.ActorInfo.BelowParty ? (int) RenderLayer.BelowPartyActor : 
-                (int) RenderLayer.AbovePartyActor;
+            var layer = actor.ActorInfo.BelowParty ? RenderLayer.BelowPartyActor : RenderLayer.AbovePartyActor;
 
             // Calculate interpolated position.
             float interpolatedX = actor.LastX + (actor.XLocation - actor.LastX) * actor.MovementProgress;
